mongoDB logic helpers
Still some bugs, but good for now

+++ mongoHelper.js
@@ -1,5 +1,6 @@
 const mongoose = require("mongoose")
 const uuidv4 = require('uuid').v4
+const helper = require("./helper")
 
 const userDataSchema = new mongoose.Schema({
     username: String,
@@ -20,13 +21,9 @@ const anonymousCartSchema = new mongoose.Schema({
 const AnonModel = new mongoose.model("AnonCart", anonymousCartSchema)
 
 const ratingsAndReviewsSchema = new mongoose.Schema({
-    allReviews: [
-        { 
-            productId: Number, 
-            reviewArray: [String],
-            ratingArray: [Number]
-        }
-    ],
+    productId: Number, 
+    reviewArray: [String],
+    ratingArray: [Number]
 })
 const FeedbackModel = new mongoose.model("feedback", ratingsAndReviewsSchema)
 
@@ -47,39 +44,74 @@ function areIdentical(oldData, newData){
     return allMatch
 }
 
-async function getMongoCart (_id) {
-    let user = await UserModel.findById( { _id } )
-    return user.shoppingCart
+async function getAllReviewsForProduct (productObj) {
+    const {productId} = productObj
+    console.log(`Product id is: [${productId}]`)
+    const feedback = await FeedbackModel.findOne({ productId })
+    console.log(`Feedback is: [${feedback}]`)
+    return feedback
 }
 
-async function getUser (idOrAnonCookie) {
-    const { _id, temporaryAnonCookie } = idOrAnonCookie
+async function getUser (idOrAnonCookieOrUsername) {
+    const { _id, temporaryAnonCookie, username } = idOrAnonCookieOrUsername
     let user = null
-    if (temporaryAnonCookie === undefined) {
-        user = await UserModel.findById( { _id } )
-    } else {
+
+    if (username !== undefined) {
+        user = await UserModel.findOne({ username: username })
+    } else if (temporaryAnonCookie !== undefined) {
         user = await AnonModel.findOne( { temporaryAnonCookie })
+    } else if (_id !== undefined){
+        user = await UserModel.findById( { id: _id } )
     }
     return user
 }
 
-async function addItemToUserReview (newReview, idOrAnonCookie) {
-    let user = await getUser(idOrAnonCookie)
+async function addReviewToMainArray (newReview) {
+    const {productId, review, rating} = newReview
+    const allReviewsForProduct = await getAllReviewsForProduct({ productId })
+    if (allReviewsForProduct === null){
+        const dataObject = {
+            productId: productId,
+            reviewArray: (review === undefined || review === null) ? [] : [ review ],
+            ratingArray: (rating === undefined || rating === null) ? [] : [ rating ]
+        }
+        await createAndSaveRecord(FeedbackModel, dataObject)
+    } else {
+        if (rating !== undefined && rating !== null) {
+            allReviewsForProduct.ratingArray.push(rating)
+        }
+        if (rating !== undefined && rating !== null) {
+            allReviewsForProduct.reviewArray.push(review)
+        }
+        await allReviewsForProduct.save()
+        return
+    }
+}
+
+async function addItemToUserReview (newReview, idOrAnonCookie) {    
+    const user = await getUser({ username: idOrAnonCookie })
     let tempArr = user.reviews
+    const {productId, review, rating} = newReview
     let indexOfExistingReview = tempArr.findIndex((item) => {
-        if (item.productId === newReview.productId) {
+        if (item.productId === productId) {
             return item
         }
     })
     if (indexOfExistingReview === -1) {
-        user.reviews.push(newReview)
-    } else {
-        Object.entries(newReview).forEach(([key, value]) => {
-            user.reviews[indexOfExistingReview][key] = value
+        user.reviews.push({
+            productId: productId,
+            review: review === undefined ? null : review,
+            rating: rating === undefined ? null : rating
         })
+        await addReviewToMainArray(newReview)
+        await user.save()
+        return true
+    } else {
+        user.reviews.rating = rating
+        await user.save()
+        await addReviewToMainArray(newReview)
+        return false
     }
-    await user.save()
-    return user
 }
 
 async function addItemToMongoCart (newProduct, idOrAnonCookie) {
@@ -87,40 +119,79 @@ async function addItemToMongoCart (newProduct, idOrAnonCookie) {
     let tempArr = user.shoppingCart
 
     const indexOfIdenticalMatch = tempArr.findIndex((item) => {
-        if (areIdentical(item, newProduct)) { 
-            return item
-        }
+        if (item.details.id === newProduct.details.id)
+            if (areIdentical(item.userSelectedParameters, newProduct.userSelectedParameters)) { 
+                return item
+            }
     })
     if( indexOfIdenticalMatch === -1) {
         tempArr.push(newProduct)
+        user.shoppingCart = tempArr
     } else {
         tempArr[indexOfIdenticalMatch].amount += newProduct.amount
+        user.shoppingCart[indexOfIdenticalMatch] = tempArr[indexOfIdenticalMatch]
     }
-
-    user.shoppingCart = tempArr
     await user.save()
     return user.shoppingCart
 }
 
-async function editItemInMongoCart (indexToChange, productWithDifferentChoices, idOrAnonCookie) {
+async function editItemInMongoCart (dataObject, idOrAnonCookie) {
     let user = await getUser(idOrAnonCookie)
-    console.log(user)
-    let tempArr = user.shoppingCart.map((item, index) => {
-        return index === indexToChange ? productWithDifferentChoices : item
-    })
-    console.log(tempArr)
+    const { productId, oldUserChoices, newUserChoices, index, amount } = dataObject
+    if(index < 0 || index >= user.shoppingCart.length){
+        return user.shoppingCart
+    }
+    let tempArr = user.shoppingCart
 
-    user.shoppingCart = tempArr
-    await user.save()
-    return user.shoppingCart
+    // setting amount to different value, on source item
+    if (areIdentical(oldUserChoices, newUserChoices)) {
+        user.shoppingCart[index].amount = amount
+        await user.save()
+        return user.shoppingCart
+    } else {
+        const indexOfIdenticalMatch = tempArr.findIndex((item) => {
+            if (item.details.id === productId)
+                if (areIdentical(item.userSelectedParameters, newUserChoices)) { 
+                    return item
+                }
+        })
+        const itemExistsInAnotherIndex = indexOfIdenticalMatch !== -1
+        if( itemExistsInAnotherIndex ) {
+            // incrementing amount of a DIFFERENT ITEM than source item
+            tempArr[indexOfIdenticalMatch].amount += amount
+        } else {
+            // CREATING NEW ITEM, because sourceItem and allother did not match with parameters
+            const dataObject = { ...helper.getProductFromProductDatabase(null, productId), userSelectedParameters: newUserChoices, amount: amount }
+            tempArr.push(dataObject)
+        }
+        // DELETE SOURCE ITEM, because we moved that data into a different object now.
+        tempArr = tempArr.filter((item, filterIndex) => {
+            if ( filterIndex !== index ) {
+                return item
+            }
+        })
+        user.shoppingCart = tempArr
+        await user.save()
+        return user.shoppingCart
+    }
 }
 
 async function deleteItemInMongoCart (indexToDelete, idOrAnonCookie) {
     let user = await getUser(idOrAnonCookie)
+    if(indexToDelete < 0 || indexToDelete >= user.shoppingCart.length){
+        return user.shoppingCart
+    }
     let tempArr = user.shoppingCart.filter((item, index) => {
         return index !== indexToDelete && item
     })
     user.shoppingCart = tempArr
+    await user.save()
+    return user.shoppingCart
+}
+
+async function clearCartInMongo (idOrAnonCookie) {
+    let user = await getUser(idOrAnonCookie)
+    user.shoppingCart = []
     await user.save()
     return user.shoppingCart
 }
@@ -134,9 +205,18 @@ async function createAnon () {
     return anon
 }
 
+async function deleteAnon (cookie) {
+    const { temporaryAnonCookie } = cookie
+    if(temporaryAnonCookie === null)
+        return null
+    
+    const result = await AnonModel.deleteOne({ temporaryAnonCookie })
+    return result.deletedCount === 1
+}
+
 async function createUserAndDeleteAnon (userCredentials, temporaryAnonCookie) {
-    await AnonModel.deleteOne({ temporaryAnonCookie })
     const { username, password } = userCredentials
+    AnonModel.deleteOne({ temporaryAnonCookie })
     const userRecord = await createAndSaveRecord(UserModel, {
         username, 
         password,
@@ -154,4 +234,5 @@ async function createAndSaveRecord (mongoModel, newParameters) {
 
 module.exports = { createAndSaveRecord, createUserAndDeleteAnon, createAnon, 
     deleteItemInMongoCart, editItemInMongoCart, addItemToMongoCart,
-    getMongoCart, connectToDatabase, getUser, addItemToUserReview }
+    connectToDatabase, getUser, addItemToUserReview, getAllReviewsForProduct,
+    clearCartInMongo, deleteAnon }
